% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluey_colors.R
\name{bluey_colors}
\alias{bluey_colors}
\title{Available Bluey Color Palettes}
\usage{
bluey_colors
}
\value{
A named \code{list} of character vectors, where each vector contains
hex color codes corresponding to a Bluey character palette
(\code{bluey}, \code{chilli}, \code{heeler}, and \code{socks}).
}
\description{
Use \code{\link{bluey_palette}} to extract palettes of the desired length
}
