% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluey_colors.R
\name{bluey_ggplot2_scales}
\alias{bluey_ggplot2_scales}
\alias{scale_color_bluey}
\alias{scale_fill_bluey}
\alias{scale_color_bluey_c}
\alias{scale_fill_bluey_c}
\title{Bluey ggplot 2 color scales}
\usage{
scale_color_bluey(option = "bluey", direction = 1, ...)

scale_fill_bluey(option = "bluey", direction = 1, ...)

scale_color_bluey_c(option = "bluey", direction = 1, ...)

scale_fill_bluey_c(option = "bluey", direction = 1, ...)
}
\arguments{
\item{option}{The name of the color palette to use. See \code{\link{bluey_palette}} for available options}

\item{direction}{If greater than or equal to 0, it will return the standard direction; otherwise it will reverse the direction of the palette}

\item{...}{Additional arguments passed to ggplot2 scale functions}
}
\value{
A \code{ScaleDiscrete} object that can be added to another \code{ggplot} object
}
\description{
ggplot2 fill and color scales derived from Bluey characters
\itemize{
\item \code{scale_color_bluey()} and \code{scale_fill_bluey()} provide discrete fill functions
\item \code{scale_color_bluey_c()} and \code{scale_fill_bluey_c()} provide continuous fill functions`
}
}
\examples{
library(ggplot2)
tmp <- data.frame(x = rnorm(100), y = rnorm(100), z = rep(c("a", "b", "c", "d"), 25))

# using a discrete scale
ggplot(tmp, aes(x, y, color = z)) +
  geom_point() +
  scale_color_bluey()

# changing the palette used when creating a discrete scale
ggplot(tmp, aes(x, y, color = z)) +
  geom_point() +
  scale_color_bluey(option = "socks")

}
