% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluey_colors.R
\docType{data}
\name{bluey_colors}
\alias{bluey_colors}
\title{Available Bluey Color Palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
bluey_colors
}
\description{
Use \code{\link{bluey_palette}} to extract palettes of the desired length
}
\keyword{datasets}
