% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluey_colors.R
\name{bluey_palette}
\alias{bluey_palette}
\title{Bluey Color Palette Constructor}
\usage{
bluey_palette(option = "bluey", n = NULL)
}
\arguments{
\item{option}{The name of the color palette to use. Current options are 'bluey', 'chilli', 'heeler', and 'socks'}

\item{n}{The number of colors to return. Currently, all scales are limited to 5 colors}
}
\description{
Extract a color palette of the desired length
}
